+++ src/main/ts/components/Editor.tsx
@@ -41,9 +41,6 @@ export interface IProps {
 
 export interface IAllProps extends Partial<IProps>, Partial<IEvents> { }
 
-const changeEvents = () => getTinymce()?.Env?.browser?.isIE() ? 'change keyup compositionend setcontent' : 'change input compositionend setcontent';
-const beforeInputEvent = () => isBeforeInputEventAvailable() ? 'beforeinput SelectionChange' : 'SelectionChange';
-
 export class Editor extends React.Component<IAllProps> {
   public static propTypes: IEditorPropTypes = EditorPropTypes;
 
@@ -67,6 +64,10 @@ export class Editor extends React.Component<IAllProps> {
     this.elementRef = React.createRef<HTMLElement>();
     this.inline = this.props.inline ?? this.props.init?.inline ?? false;
     this.boundHandlers = {};
+  }
+
+  private get view() {
+    return this.elementRef.current?.ownerDocument.defaultView ?? window;
   }
 
   public componentDidUpdate(prevProps: Partial<IAllProps>) {
@@ -121,7 +122,7 @@ export class Editor extends React.Component<IAllProps> {
   }
 
   public componentDidMount() {
-    if (getTinymce() !== null) {
+    if (getTinymce(this.view) !== null) {
       this.initialise();
     } else if (this.elementRef.current && this.elementRef.current.ownerDocument) {
       ScriptLoader.load(
@@ -138,8 +139,8 @@ export class Editor extends React.Component<IAllProps> {
   public componentWillUnmount() {
     const editor = this.editor;
     if (editor) {
-      editor.off(changeEvents(), this.handleEditorChange);
-      editor.off(beforeInputEvent(), this.handleBeforeInput);
+      editor.off(this.changeEvents(), this.handleEditorChange);
+      editor.off(this.beforeInputEvent(), this.handleBeforeInput);
       editor.off('keypress', this.handleEditorChangeSpecial);
       editor.off('keydown', this.handleBeforeInputSpecial);
       editor.off('NewBlock', this.handleEditorChange);
@@ -154,6 +155,15 @@ export class Editor extends React.Component<IAllProps> {
 
   public render() {
     return this.inline ? this.renderInline() : this.renderIframe();
+  }
+
+  private changeEvents() {
+    const isIE = getTinymce(this.view)?.Env?.browser?.isIE();
+    return isIE ? 'change keyup compositionend setcontent' : 'change input compositionend setcontent';
+  }
+
+  private beforeInputEvent() {
+    return isBeforeInputEventAvailable() ? 'beforeinput SelectionChange' : 'SelectionChange';
   }
 
   private renderInline() {
@@ -203,14 +213,14 @@ export class Editor extends React.Component<IAllProps> {
       const wasControlled = isValueControlled(prevProps);
       const nowControlled = isValueControlled(this.props);
       if (!wasControlled && nowControlled) {
-        this.editor.on(changeEvents(), this.handleEditorChange);
-        this.editor.on(beforeInputEvent(), this.handleBeforeInput);
+        this.editor.on(this.changeEvents(), this.handleEditorChange);
+        this.editor.on(this.beforeInputEvent(), this.handleBeforeInput);
         this.editor.on('keydown', this.handleBeforeInputSpecial);
         this.editor.on('keyup', this.handleEditorChangeSpecial);
         this.editor.on('NewBlock', this.handleEditorChange);
       } else if (wasControlled && !nowControlled) {
-        this.editor.off(changeEvents(), this.handleEditorChange);
-        this.editor.off(beforeInputEvent(), this.handleBeforeInput);
+        this.editor.off(this.changeEvents(), this.handleEditorChange);
+        this.editor.off(this.beforeInputEvent(), this.handleBeforeInput);
         this.editor.off('keydown', this.handleBeforeInputSpecial);
         this.editor.off('keyup', this.handleEditorChangeSpecial);
         this.editor.off('NewBlock', this.handleEditorChange);
@@ -238,7 +248,7 @@ export class Editor extends React.Component<IAllProps> {
 
   private handleBeforeInput = (_evt: EditorEvent<unknown>) => {
     if (this.props.value !== undefined && this.props.value === this.currentContent && this.editor) {
-      if (!this.inline || this.editor.hasFocus) {
+      if (!this.inline || this.editor.hasFocus()) {
         try {
           // getBookmark throws exceptions when the editor has not been focused
           // possibly only in inline mode but I'm not taking chances
@@ -305,7 +315,7 @@ export class Editor extends React.Component<IAllProps> {
       return;
     }
 
-    const tinymce = getTinymce();
+    const tinymce = getTinymce(this.view);
     if (!tinymce) {
       throw new Error('tinymce should have been loaded into global scope');
     }